+++ documentation/FuriHalDebuging.md
@@ -24,3 +24,7 @@ There are 2 signals that will be exposed to external GPIO pins:
 
 - `WFI`     - `PB2` - Light sleep (wait for interrupt) used. Basically this is lightest and most non-breaking things power save mode. All function and debug should work correctly in this mode.
 - `STOP`    - `PC3` - STOP mode used. Platform deep sleep mode. Extremely fragile mode where most of the silicon is disabled or in unusable state. Debugging MCU in this mode is nearly impossible.
+
+## FuriHalSD
+
+`--extra-define=FURI_HAL_SD_SPI_DEBUG` enables SD card SPI bus logging.
