+++ firmware/targets/f7/fatfs/user_diskio.h
@@ -4,7 +4,6 @@
 extern "C" {
 #endif
 
-#include "sd_spi_io.h"
 #include "fatfs/ff_gen_drv.h"
 
 extern Diskio_drvTypeDef sd_fatfs_driver;
