+++ firmware/targets/f7/src/update.c
@@ -23,8 +23,8 @@ static FATFS* pfs = NULL;
     }
 
 static bool flipper_update_mount_sd() {
-    for(int i = 0; i < sd_max_mount_retry_count(); ++i) {
-        if(sd_init((i % 2) == 0) != SdSpiStatusOK) {
+    for(int i = 0; i < furi_hal_sd_max_mount_retry_count(); ++i) {
+        if(furi_hal_sd_init((i % 2) == 0) != FuriStatusOk) {
             /* Next attempt will be without card reset, let it settle */
             furi_delay_ms(1000);
             continue;
@@ -51,7 +51,7 @@ static bool flipper_update_init() {
     furi_hal_spi_config_init();
 
     fatfs_init();
-    if(!hal_sd_detect()) {
+    if(!furi_hal_sd_is_present()) {
         return false;
     }
